+++ access-control-sequelize/lib/index.ts
@@ -1,2 +1,3 @@
 export * from "./hooks/sequelize-hooks.module";
 export * from "./models/sequelize-models.service";
+export * from "./repositories/repositories";